07-langgraph/pyproject.toml
[project]
name = "1-one-node"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.21",
    "langgraph>=0.4.8",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "wolframalpha>=5.1.3",
    "yfinance>=0.2.61",
]
